Move singleton tests
